+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 + Add callback and default implementation `Recode.Task.new_issue/1` and 
   `Recode.Task.new_issue/2` 
 + Rename `Recode.Task.TestFileExt` to `Recode.Task.TestFile`
-+ Updte `mix recode.update.config` to remove deprecated tasks
++ Update `mix recode.update.config` to remove deprecated tasks
 
 ## 0.7.3 - 2024/07/25