+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 + Update `mix recode.update.config` to remove deprecated tasks.
 + Add a manifest file to store information about the files processed.
 + Add silent mode with flag `--silent`.
++ Fix `Recode.Task.LocalsWithoutParens`.
 
 ## 0.7.3 - 2024/07/25